Implement additional form elements and sidebar navigation
This commit adds new form fields for "agenzia", "anno", "data_pagamento", and "fine_validita". It also rolls out new sidebar navigation for "Veicolo" and "Stazionamento Prolungato", as well as their respective "Scadenze" sub-items. The sidebar for "Alerts" is updated and tidied up. The changes will facilitate better navigation and usage for users.

+++ resources/views/sidebar/decorazione/menu.blade.php
@@ -42,30 +42,34 @@
 			<!--end:Menu link-->
 		</div>
 		<!--end:Menu item-->
-		<!--begin:Menu item-->
-		<div class="menu-item">
-			<!--begin:Menu link-->
-			<a class="menu-link" href="/update_decorazione">
+		@if(request()->is('update_decorazione/*'))
+			<!--begin:Menu item-->
+			<div class="menu-item">
+				<!--begin:Menu link-->
+				<a class="menu-link" href="/update_decorazione">
 									<span class="menu-bullet">
 											<span class="bullet bullet-dot"></span>
 									</span>
-				<span class="menu-title">Aggiorna</span>
-			</a>
-			<!--end:Menu link-->
-		</div>
-		<!--end:Menu item-->
-		<!--begin:Menu item-->
-		<div class="menu-item">
-			<!--begin:Menu link-->
-			<a class="menu-link" href="/delete_decorazione">
+					<span class="menu-title">Aggiorna</span>
+				</a>
+				<!--end:Menu link-->
+			</div>
+			<!--end:Menu item-->
+		@endif
+		@if(request()->is('delete_decorazione/*'))
+			<!--begin:Menu item-->
+			<div class="menu-item">
+				<!--begin:Menu link-->
+				<a class="menu-link" href="/delete_decorazione">
 									<span class="menu-bullet">
 											<span class="bullet bullet-dot"></span>
 									</span>
-				<span class="menu-title">Cancella</span>
-			</a>
-			<!--end:Menu link-->
-		</div>
-		<!--end:Menu item-->
+					<span class="menu-title">Cancella</span>
+				</a>
+				<!--end:Menu link-->
+			</div>
+			<!--end:Menu item-->
+		@endif
 	</div>
 	<!--end:Menu sub-->
 </div>
